% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremeid.R
\name{extremeid}
\alias{extremeid}
\title{Summarize Note Character Length}
\usage{
extremeid(dataset, extreme, clean_notes, pageid, group_list = NA)
}
\arguments{
\item{dataset}{data frame of notes}

\item{extreme}{number of standard deviations above the mean character length that defines an extreme value}

\item{clean_notes}{column name of clean notes to provide summary values for}

\item{pageid}{column name for page number}

\item{group_list}{list of variables to group by for cleaned notes}
}
\value{
a data frame
}
\description{
Summarize Note Character Length
}
\examples{
test_dataset <- data.frame(ID=c("1","1","2","2","1"),
Notes=c("The","cat","The","dog","ran"),
Page=c(1,2,1,2,3))
extremeid(dataset=test_dataset,clean_notes="Notes",extreme=2,pageid="Page")
}
