% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstnchar.R
\name{firstnchar}
\alias{firstnchar}
\title{First N Character Note Cleaning}
\usage{
firstnchar(dataset, notes, char_diff, identifier, pageid)
}
\arguments{
\item{dataset}{the dataset containing the notes}

\item{notes}{the column name for the notes}

\item{char_diff}{allowable character difference for removing notes}

\item{identifier}{column name for uniquely identifying identification}

\item{pageid}{column name for page number}
}
\value{
a data frame
}
\description{
First N Character Note Cleaning
}
\examples{
test_dataset <- data.frame(ID=c("1","1","2","2","1", "3", "3"),
Notes=c("The","The cat","The","The dog","The cat ran", "the chicken was chased", "The goat chased the chicken"),
Page=c(1,2,1,2,3,1,2))
firstnchar(dataset=test_dataset,notes="Notes",char_diff=3,identifier="ID",pageid="Page")
}
