% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcsclean_hybrid.R
\name{lcsclean_hybrid}
\alias{lcsclean_hybrid}
\title{Longest Common Substring Note Cleaning for Hybrid Method}
\usage{
lcsclean_hybrid(dataset, notes, propor, identifier, pageid, toclean)
}
\arguments{
\item{dataset}{the dataset containing the notes}

\item{notes}{the column name for the notes}

\item{propor}{minimum necessary of matching proportion of previous notes for removal}

\item{identifier}{column name for uniquely identifying identification}

\item{pageid}{column name for page number}

\item{toclean}{column name for notes to clean}
}
\value{
a data frame
}
\description{
Longest Common Substring Note Cleaning for Hybrid Method
}
