% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validation_dataset}
\alias{validation_dataset}
\title{Hybrid Method Validation Dataset}
\format{
\subsection{\code{validation_dataset}}{

A data frame with 561 rows and 4 columns:
\describe{
\item{clean_prints}{Identifier}
\item{page_count}{page number}
\item{notes}{The uncleaned notes}
\item{corrected_notes}{hand cleaned notes to demonstrate correct format}

}
}
}
\source{
Jury Perception Study
}
\usage{
validation_dataset
}
\description{
A subset of participant responses from Jury Perception study
}
\keyword{datasets}
