% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremeid.R
\name{extremeid}
\alias{extremeid}
\title{Summarize Note Character Length}
\usage{
extremeid(dataset, extreme, clean_notes, pageid, group_list = NA)
}
\arguments{
\item{dataset}{data frame of notes}

\item{extreme}{number of standard deviations above the mean character length that defines an extreme value}

\item{clean_notes}{column name of clean notes to provide summary values for}

\item{pageid}{column name for page number}

\item{group_list}{list of variables to group by for cleaned notes}
}
\value{
a data frame
}
\description{
Use to identify extreme values, based on cleaned note length, in a dataset.
Can be used after applying \code{\link[=firstnchar]{firstnchar()}}
}
\examples{
test_dataset <- data.frame(ID=c("1","1","2","2","1", "3","3"),
Notes=c("The","The cat","The","The dog","The cat ran",
"the chicken was chased", "The goat chased the chicken"),
Page=c(1,2,1,2,3,1,2))
cleaned_dataset<-
firstnchar(dataset=test_dataset,notes="Notes",char_diff=3,
identifier="ID",pageid="Page")
extremeid(dataset=cleaned_dataset,clean_notes="page_notes",extreme=2,pageid="Page")
}
