% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcsclean_hybrid.R
\name{lcsclean_hybrid}
\alias{lcsclean_hybrid}
\title{Longest Common Substring Note Cleaning for Hybrid Method}
\usage{
lcsclean_hybrid(dataset, notes, propor, identifier, pageid, toclean)
}
\arguments{
\item{dataset}{the dataset containing the notes}

\item{notes}{the column name for the notes}

\item{propor}{minimum necessary of matching proportion of previous notes for removal}

\item{identifier}{column name for uniquely identifying identification}

\item{pageid}{column name for page number}

\item{toclean}{column name for identifying column of notes to clean (TRUE/FALSE)}
}
\value{
a data frame
}
\description{
This function is used to apply the longest common substring method to extreme values in a dataset.
To be used after applying \code{\link[=firstnchar]{firstnchar()}} and \code{\link[=extremeid]{extremeid()}}. Dataset should have
a "page_notes" column corresponding to the cleaned notes outcome from \code{\link[=firstnchar]{firstnchar()}}.
}
\examples{
test_dataset <- data.frame(ID=c("1","1","2","2","1", "3","3"),
Notes=c("The","The cat","The","The dog","The cat ran",
"the chicken was chased", "The goat chased the chicken"),
Page=c(1,2,1,2,3,1,2), cleaning = c(FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, TRUE),
page_notes = c("The","The cat","The","The dog","The cat ran",
"the chicken was chased", "The goat chased the chicken"))
lcsclean_hybrid(test_dataset,"Notes",0.5,"ID","Page", "cleaning")
}
